feat(suite): simplified ConnectionGlobalModal and ThpModals
