+++ packages/suite/src/components/connection/ConnectionGlobalModal.tsx
@@ -1,56 +1,21 @@
 import { selectThpStep } from '@suite-common/thp';
-import { selectSelectedFirstThpDevice } from '@suite-common/wallet-core';
-import { exhaustive } from '@trezor/type-utils';
 
 import { selectIsConnectionModalOpen } from 'src/actions/device/deviceSelectors';
 import { setConnectionModal } from 'src/actions/device/deviceSlice';
 import { useDispatch, useSelector } from 'src/hooks/suite';
 
 import { ConnectDeviceGlobalModal } from './ConnectDeviceGlobalModal';
-import { ThpAutoconnectInfoModal } from './thp/ThpAutoconnectInfoModal';
-import { ThpAutoconnectionModal } from './thp/ThpAutoconnectionModal';
-import { ThpConnectionModal } from './thp/ThpConnectionModal';
-import { ThpPairingFailedModal } from './thp/ThpPairingFailedModal';
-import { ThpPairingPinEntryModal } from '../suite/modals/ReduxModal/DeviceContextModal/ThpPairingPinEntryModal';
 
-type ConnectionGlobalModalProps = {
-    /** @deprecated Should be removed once we get rid of onboarding fullscreen app logic. */
-    showThpModals?: boolean;
-};
-
-export const ConnectionGlobalModal = ({ showThpModals = true }: ConnectionGlobalModalProps) => {
+export const ConnectionGlobalModalManager = () => {
     const dispatch = useDispatch();
     const isConnectDeviceModalOpen = useSelector(selectIsConnectionModalOpen);
-    const device = useSelector(selectSelectedFirstThpDevice);
     const thpStep = useSelector(selectThpStep);
 
     const closeConnectionModal = () => {
         dispatch(setConnectionModal(false));
     };
 
-    // handle THP modals first if we have a device and thpStep
-    if (device !== undefined && thpStep !== null && showThpModals) {
-        switch (thpStep) {
-            case 'BeforeConnectionInfo':
-                return null;
-            case 'ConfirmConnectionBeforePairing':
-                return <ThpConnectionModal device={device} />;
-            case 'ConfirmOnlyConnection':
-                return <ThpConnectionModal device={device} />;
-            case 'CodeEntry':
-                return <ThpPairingPinEntryModal />;
-            case 'CodeInvalid':
-                return <ThpPairingFailedModal />;
-            case 'AutoconnectInfo':
-                return <ThpAutoconnectInfoModal device={device} />;
-            case 'Autoconnect':
-                return <ThpAutoconnectionModal device={device} />;
-            default:
-                return exhaustive(thpStep);
-        }
-    }
-
-    if (!isConnectDeviceModalOpen) return null;
+    if (!isConnectDeviceModalOpen || thpStep !== null) return null;
 
     return <ConnectDeviceGlobalModal onCancel={closeConnectionModal} />;
 };
